requirements.txt
python-telegram-bot>=22.0
python-jobspy>=1.1.0
pydantic>=2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0
weasyprint>=62.0
python-docx>=1.1.0
jinja2>=3.1.0
python-dotenv>=1.0.0
# Optional: only needed if using anthropic as LLM provider
# anthropic>=0.42.0
